Refactor package into subpackages
